Move verifier tests to core verifier tests
